Interface refinements for iozeug, threadingzeug and reflectionzeug

+++ source/threadingzeug/include/threadingzeug/parallelfor.h
@@ -8,9 +8,9 @@
 namespace threadingzeug
 {
 
-THREADINGZEUG_API void forEach(int start, int end, std::function<void(int i)> callback, bool parallelize = true);
+THREADINGZEUG_API void forEach(size_t start, size_t end, std::function<void(size_t i)> callback, bool parallelize = true);
 
-using ForSignature = void (*)(int, int, std::function<void(int i)>);
+using ForSignature = void (*)(size_t, size_t, std::function<void(size_t i)>);
 
 template <typename T, typename U>
 ForSignature selectParallelization(const T & value, const U & threshold);
@@ -23,7 +23,7 @@ void parallelFor(const std::vector<T>& elements, std::function<void(const T& ele
 template<typename T>
 void parallelFor(std::vector<T>& elements, std::function<void(T& element)> callback);
 
-THREADINGZEUG_API void parallelFor(int start, int end, std::function<void(int i)> callback);
+THREADINGZEUG_API void parallelFor(size_t start, size_t end, std::function<void(size_t i)> callback);
 
 
 template<typename T>
@@ -32,7 +32,7 @@ void sequentialFor(const std::vector<T>& elements, std::function<void(const T& e
 template<typename T>
 void sequentialFor(std::vector<T>& elements, std::function<void(T& element)> callback);
 
-THREADINGZEUG_API void sequentialFor(int start, int end, std::function<void(int i)> callback);
+THREADINGZEUG_API void sequentialFor(size_t start, size_t end, std::function<void(size_t i)> callback);
 
 } // namespace threadingzeug
